@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "com.squizzard.MisriCalendar"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         // used by Room, to test migrations
         javaCompileOptions {
             annotationProcessorOptions {
@@ -38,12 +38,10 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'com.j256.ormlite:ormlite-android:4.48'
     implementation 'com.facebook.stetho:stetho:1.5.0'
 
-    def room_version = "1.1.1"
-
-    implementation "android.arch.persistence.room:runtime:$room_version"
-    annotationProcessor "android.arch.persistence.room:compiler:$room_version"
+    implementation 'androidx.room:room-runtime:2.2.3'
+    annotationProcessor 'androidx.room:room-compiler:2.2.3'
 }
