@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import com.squizzard.util.DateUtil;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
