@@ -1,6 +1,6 @@
 package com.squizzard.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Map;
 import java.util.TreeMap;
