@@ -1,14 +1,12 @@
 package com.squizzard.Database;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
 
-import com.j256.ormlite.stmt.DeleteBuilder;
 import com.squizzard.Reminder.Reminder;
 
-import java.sql.SQLException;
 import java.util.ArrayList;
 
 @Dao
